+++ crates/bcr-ebill-core/src/util/mod.rs
@@ -1,3 +1,4 @@
+pub mod borsh;
 pub mod crypto;
 pub mod currency;
 pub mod date;